新数据：优化工时导入

+++ src/main/java/cn/com/tarotframework/server/oak/dto/ProjectHour.java
@@ -2,8 +2,10 @@ package cn.com.tarotframework.server.oak.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -23,11 +25,13 @@ public class ProjectHour {
 
     private String projectName;
 
-    private Double totalHour;
+    //当前天参与所有项目总工时
+    private BigDecimal totalHour;
 
     // 填报日期
     private LocalDate fillDate;
 
+    // 创建日期
     private LocalDateTime createTime;
 
     List<ProjectHourDetail> projectHourDetails;
